Feat(LOCAT-BE-18): Move WebhookRequest package path
